Refactor image optimization stack into a Construct.
